firepower_import_csv/requirements.txt
# List third-party dependencies here, separated by newlines.
# All dependencies must be version-pinned, eg. requests==1.2.0
# See: https://pip.pypa.io/en/stable/user_guide/#requirements-files
pyOpenSSL==19.1.0
beautifulsoup4==4.8.0
paramiko==2.7.1
scp==0.13.2
